+++ test/unit/client.test.ts
@@ -80,7 +80,6 @@ describe("Actos Client", () => {
     expect(client.search).toBeDefined();
     expect(client.votes).toBeDefined();
     expect(client.saves).toBeDefined();
-    expect(client.uploads).toBeDefined();
     expect(client.reports).toBeDefined();
     expect(client.admin).toBeDefined();
     expect(client.meta).toBeDefined();